+++ src/main/java/io/bootique/liquibase/command/ChangelogSyncCommand.java
@@ -12,6 +12,9 @@ import liquibase.LabelExpression;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/**
+ * @since 0.11
+ */
 public class ChangelogSyncCommand extends CommandWithMetadata {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ChangelogSyncCommand.class);