+++ src/main/java/io/bootique/liquibase/command/ClearCheckSumsCommand.java
@@ -10,6 +10,9 @@ import io.bootique.liquibase.LiquibaseRunner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/**
+ * @since 0.11
+ */
 public class ClearCheckSumsCommand extends CommandWithMetadata {
     private static final Logger LOGGER = LoggerFactory.getLogger(ClearCheckSumsCommand.class);