+++ src/main/java/io/bootique/liquibase/command/UpdateCommand.java
@@ -12,6 +12,7 @@ import liquibase.LabelExpression;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+
 public class UpdateCommand extends CommandWithMetadata {
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(UpdateCommand.class);