+++ src/main/java/io/bootique/liquibase/LiquibaseModule.java
@@ -10,15 +10,12 @@ import io.bootique.config.ConfigurationFactory;
 import io.bootique.jdbc.DataSourceFactory;
 import io.bootique.liquibase.annotation.ChangeLogs;
 import io.bootique.liquibase.command.ChangelogSyncCommand;
-import io.bootique.liquibase.command.ChangelogSyncSqlCommand;
 import io.bootique.liquibase.command.ClearCheckSumsCommand;
 import io.bootique.liquibase.command.UpdateCommand;
 import io.bootique.liquibase.command.ValidateCommand;
 import io.bootique.resource.ResourceFactory;
 
 import java.util.Set;
-
-import static java.util.Arrays.asList;
 
 public class LiquibaseModule extends ConfigModule {
 
@@ -29,21 +26,39 @@ public class LiquibaseModule extends ConfigModule {
     public LiquibaseModule() {
     }
 
+    /**
+     * @param binder DI binder passed to the Module that invokes this method.
+     * @return an instance of {@link LiquibaseModuleExtender} that can be used to load custom extensions.
+     * @since 0.12
+     */
+    public static LiquibaseModuleExtender extend(Binder binder) {
+        return new LiquibaseModuleExtender(binder);
+    }
+
+    /**
+     * Returns a Guice {@link Multibinder} to add Liquibase change logs
+     *
+     * @param binder DI binder passed to the Module that invokes this method.
+     * @return returns a {@link Multibinder} for Liquibase change logs.
+     * @deprecated since 0.12 call {@link #extend(Binder)} and then call
+     * {@link LiquibaseModuleExtender#addChangeLog(ResourceFactory)}.
+     */
+    @Deprecated
     public static Multibinder<ResourceFactory> contributeChangeLogs(Binder binder) {
         return Multibinder.newSetBinder(binder, ResourceFactory.class, ChangeLogs.class);
     }
 
     @Override
     public void configure(Binder binder) {
-        asList(
-                UpdateCommand.class,
-                ValidateCommand.class,
-                ChangelogSyncCommand.class,
-                ClearCheckSumsCommand.class,
-                ChangelogSyncSqlCommand.class
-        ).forEach(command ->
-                BQCoreModule.contributeCommands(binder).addBinding().to(command).in(Singleton.class));
-        contributeChangeLogs(binder);
+
+        LiquibaseModule.extend(binder).initAllExtensions();
+
+        BQCoreModule.extend(binder)
+                .addCommand(UpdateCommand.class)
+                .addCommand(ValidateCommand.class)
+                .addCommand(ChangelogSyncCommand.class)
+                .addCommand(ClearCheckSumsCommand.class)
+                .addCommand(ChangelogSyncCommand.class);
     }
 
     @Provides