+++ RELEASE-NOTES.md
@@ -1,6 +1,6 @@
 ## 0.11
 
-* #7 Replace 'changeLog' with 'changeLogs' and map Strings to ResourceFactory
+* #4 Replace 'changeLog' with 'changeLogs' and map Strings to ResourceFactory
 * #8 Upgrade to the latest Liquibase 3.5.3
 * #9 Upgrades to latest BQ, JDBC, LB